Add static checker config + small fixes (clang-tidy, clazy)

+++ include/lsl_cpp.h
@@ -790,7 +790,7 @@ private:
 
 	/// Check whether a given data length matches the number of channels; throw if not
 	void check_numchan(std::size_t N) const {
-		if (N != channel_count)
+		if (N != static_cast<std::size_t>(channel_count))
 			throw std::runtime_error("Provided element count (" + std::to_string(N) +
 									 ") does not match the stream's channel count (" +
 									 std::to_string(channel_count) + '.');
